feat:메뉴 수정 완려

+++ src/main/java/com/itwill/matzip/repository/MenuRepository.java
@@ -3,5 +3,5 @@ package com.itwill.matzip.repository;
 import com.itwill.matzip.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuRepository extends JpaRepository<Menu, Integer> {
+public interface MenuRepository extends JpaRepository<Menu, Long> {
 }
